requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
anyio==3.5.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
boto3==1.24.25
botocore==1.27.25
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.2
Django==4.0.2
dnspython==2.2.1
dynamodb==1.0.0
ecdsa==0.18.0
email-validator==1.2.1
fanstatic==1.1
fastapi==0.72.0
Flask==2.0.3
Flask-Cors==3.0.10
frozenlist==1.3.0
geckodriver-autoinstaller==0.1.0
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
h11==0.12.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.1
javascript==1!0.2.15
Jinja2==3.0.3
jmespath==1.0.1
js==1.0
Js2Py==0.71
Mako==1.2.1
MarkupSafe==2.0.1
mathx==0.2.2
multidict==6.0.1
mypy==0.961
mypy-extensions==0.4.3
numpy==1.22.2
oauth2client==4.1.3
outcome==1.1.0
pandas==1.4.0
Parlance==1.4.1
passlib==1.7.4
postgrest==0.0.1
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyDrive==1.3.1
pyjsparser==2.7.1
pymongo==4.1.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.27.1
rsa==4.8
s3transfer==0.6.0
scipy==1.9.0
selenium==3.5.0
shutilwhich==1.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
starlette==0.17.1
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.7
WebOb==1.8.7
websockets==10.1
Werkzeug==2.0.3
wsproto==1.1.0
xlrd==2.0.1
yarl==1.7.2
